Move logic to service fiel
